commit with gacp

+++ Module-07/ex02/Array.hpp
@@ -89,7 +89,7 @@ size_t Array<T>::getSize( void ) const
 }
 
 template<class T>
-const Array<T>& Array<T>::operator[](unsigned int index) const
+const Array<T>& Array<T>::operator[](size_t index) const
 {
     if (index >= _size)
 	{
@@ -99,7 +99,7 @@ const Array<T>& Array<T>::operator[](unsigned int index) const
 }
 
 template<class T>
-Array<T>& Array<T>::operator[](unsigned int index)
+Array<T>& Array<T>::operator[](size_t index)
 {
     if (index >= _size)
 	{
